feat(japp): add react-navigation
