+++ src/screen/Home/Home.tsx
@@ -1,0 +1,51 @@
+import React, { FC, useState, useEffect } from 'react';
+import { StyleSheet, Text, View } from 'react-native';
+import firebase from 'firebase/app';
+
+const Home: FC = () => {
+  // Set an initializing state whilst Firebase connects
+  const [initializing, setInitializing] = useState(true);
+  const [user, setUser] = useState<firebase.User | null>();
+
+  // Handle user state changes
+  function onAuthStateChanged(newUser: firebase.User | null) {
+    setUser(newUser);
+    if (initializing) {
+      setInitializing(false);
+    }
+  }
+
+  useEffect(() => {
+    const subscriber = firebase.auth().onAuthStateChanged(onAuthStateChanged);
+    return subscriber; // unsubscribe on unmount
+  });
+
+  if (initializing) {
+    return null;
+  }
+
+  if (!user) {
+    return (
+      <View>
+        <Text>Login</Text>
+      </View>
+    );
+  }
+
+  return (
+    <View style={styles.container}>
+      <Text>Welcome {user.email}</Text>
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+});
+
+export default Home;
